Use XInput2 to log (key|button) press/release events and scroll deltas
In the XInput example, map key/button/axis events to
key|button press/release actions and scroll events, keeping
track of the last-seen state of scroll axies
to calculate the horizontal/vertical scroll deltas.

This provides the kind of information that a higher-level
toolkit or wrapper (eg. Glutin) would need.