Uncomment None constant
I'm assuming this was commented out to avoid shadowing
`std::option::Option::None`. However, aside from the fact that this
essentially excludes a symbol from the library bindings that should be
included, this is really not necessary. There are other ways for client
code to avoid this issue, such as avoiding importing `x11::xlib::*` or
re-exporting either `std::option::Option::None` or `x11::xlib::None`
under a different name.